Forbid deleting a cluster in creating state
If create and delete playbooks (post and delete API requests) are
launched immediately after each other, resources in OpenStack will be
allocated for a cluster that no longer exists in the API because delete
is much faster. That leaves allocated resources in OS that cannot be
viewed/deleted via RHub.